\name{removeNTC-methods}
\docType{methods}
\alias{removeNTC}
\alias{removeNTC-methods}
\alias{removeNTC,ddCtExpression-method}
\alias{removeNTC,InputFrame-method}
\title{ Remove NTC samples}
\description{
  NTC stands for Non-template controls. This method remove the NTC
  samples from the input object.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ddCtExpression")}}{
  An object hat has been analyzed with the \code{ddCt} method
}

\item{\code{signature(object = "InputFrame")}}{
  An input object
}
}}
\keyword{methods}
